Add BBR protocol

+++ CA4 - Congestion Control Algorithms/src/constants.hpp
@@ -2,7 +2,7 @@
 #define CONSTANTS_HPP
 
 const int MSS = 1;
-const int RTT = 10;
+const double RTT = 10;
 const int LOSS_TRESH = 800;
 const double LOSS_SCALE = 0.015;
